third_party/WebKit/LayoutTests/http/tests/sendbeacon/beacon-cross-origin-redirect-expected.txt

PingLoader dispatched to 'http://127.0.0.1:8080/resources/redirection-response.php?status=302&target=/non-existent.php'.
CONSOLE ERROR: Redirect from 'http://127.0.0.1:8080/resources/redirection-response.php?status=302&target=/non-existent.php' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Verifying that navigator.sendBeacon() to non-CORS cross-origin redirect fails.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.sendBeacon("http://127.0.0.1:8080/resources/redirection-response.php?status=302&target=/non-existent.php", "CrossOrigin"); is true
PASS successfullyParsed is true

TEST COMPLETE
